Move createUser validation to the model method
